NEXT-25051 - Cleanup tests

+++ tests/unit/php/Core/Framework/Log/Monolog/AnnotatePackageHandlerTest.php
@@ -221,7 +221,7 @@ class TestCommand extends Command
         $testCause = new TestCause();
         $testCause->throw(new TestException('test'));
 
-        return (int) Command::SUCCESS;
+        return Command::SUCCESS;
     }
 }
